add data client

+++ omartech4j/laibican-robot/src/main/java/com/omartech/laibicanRobot/model/Article.java
@@ -5,7 +5,7 @@ package com.omartech.laibicanRobot.model;
  */
 public class Article {
 
-    private int id;
+    private long id;
     private String articleId;//
     private String title;
     private String content;
@@ -56,12 +56,11 @@ public class Article {
         this.createdAt = createdAt;
     }
 
-    public int getId() {
-
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
